requirements.txt
# Core web framework and API dependencies
streamlit>=1.40.0
requests>=2.25.0
websockets>=12.0

# Data processing and analysis
numpy>=1.20.0
pandas>=2.0.0
scipy>=1.7.0

# Computer vision and machine learning
opencv-python-headless>=4.5.0
ultralytics>=8.0.0
supervision>=0.20.0

# Image processing
pillow>=8.0.0
matplotlib>=3.5.0

# Geospatial and mapping
folium>=0.15.0
streamlit_folium>=0.20.0
earthengine-api>=0.1.400s
rasterio>=1.3.0
geemap>=0.32.0

# UI components
streamlit-option-menu>=0.3.0

# Data serialization and configuration
PyYAML>=6.0.0
toml>=0.10.0

# Async and HTTP client dependencies
aiohttp>=3.8.0